chore: update imports for moved and renamed files
