+++ src/features/i18n/LanguageSwitch/BrowserLanguageRedirect.astro
@@ -1,5 +1,5 @@
 ---
-import { locales, defaultLocale } from "src/features/i18n/config";
+import { defaultLocale, locales } from "@data/i18nConfig";
 
 const nonDefaultLocales = locales.filter((l) => l !== defaultLocale);
 ---
