+++ src/js/routes.ts
@@ -1,4 +1,4 @@
-import { defaultLocale } from "../features/i18n/config";
+import { defaultLocale } from "../data/i18nConfig";
 
 function normalizePath(path: string) {
   // 1. Force it to be a string or fallback to empty string if it's null/undefined
